feat: adds back button to onboarding screens
