+++ components/brave_wallet_ui/components/desktop/wallet-sub-view-layout/style.ts
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 export const StyledWrapper = styled('div') <{}>`
   display: flex;
   flex-direction: column;
-  align-items: center;
+  align-items: flex-start;
   justify-conent: center;
   flex: 1;
   height: 100%;
