+++ components/brave_wallet_ui/components/shared/index.ts
@@ -1,9 +1,11 @@
 import AppListItem from './app-list-item'
 import AppList from './app-list'
 import SearchBar from './search-bar'
+import BackButton from './back-button'
 
 export {
   AppListItem,
   AppList,
-  SearchBar
+  SearchBar,
+  BackButton
 }
